requirements.txt
accelerate==0.23.0
datasets==2.14.6
deepspeed==0.11.0
numpy==1.26.2
peft==0.5.0
PyYAML==6.0.1
submitit==1.5.1
torch==2.0.1
torchvision==0.15.2
transformers==4.32.1
trl==0.7.2
einops
tiktoken
matplotlib
pillow
transformers_stream_generator
wandb
